feat(graphView): able to select the graph points in normal graph mode
Enable the normal graph mode to accept the tap gesture to select two points with subsequet dialog box to enter calibration value.
This change will be revisted and updated as this change is only made to check the working of this feature.
With this change, the tap gestures in exclusive mode is affected. The fix will be in further commits.